chore: Add code-split for Git integration for packages (#39166)
## Description
EE counterpart: https://github.com/appsmithorg/appsmith-ee/pull/6032


Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.ImportExport, @tag.Git"

### :mag: Cypress test results
